+++ webserv/includes/server/Connection.hpp
@@ -6,7 +6,7 @@
 /*   By: tat-nguy <tat-nguy@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/09/25 10:22:05 by tat-nguy          #+#    #+#             */
-/*   Updated: 2025/09/26 15:58:18 by tat-nguy         ###   ########.fr       */
+/*   Updated: 2025/09/30 13:35:22 by tat-nguy         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,23 +18,35 @@
 # include "ILoop.hpp"
 # include "PollLoop.hpp"
 
+class Server;
+
 class Connection: public IEventHandler
 {
 	private:
-		PollLoop	*_loop;		// take from Server
-		int			_clientFd;	// from Listener
-		std::string	_buffer;
+		Server	*_server; // non-owning pointer back to server for marking close/mod events	
+		int		_clientFd;
+		std::string	_inBuf;
+		std::string	_outBuf;
+		bool	_hasActiveRequest;
+	
 		
 		Connection	&operator=(Connection const &rhs);
+		Connection(Connection const &src);
 	
 	public:
 		Connection(void);
-		Connection(PollLoop &loop, int cfd);
-		Connection(Connection const &src);
+		Connection(Server *server, int cfd);
+		
 		virtual ~Connection();
 
 		virtual int		getFd(void) const;
 		virtual void	handleEvent(uint32_t events);
+
+		bool	recvIntoBuffer(void);
+		bool	flushOutBuffer(void);
+
+		// for Request and Response
+		void	queueResponse(std::string const &s);
 };
 
 #endif
